refactor(modelHandlers): extract provider helpers, tighten type-only imports
Lift pure helpers out of the Google, OpenRouter, and OpenAI chat handlers
(message validation, stream aggregation, reasoning extraction) into sibling
modules with re-exports preserving the public entry points; convert
type-only value imports across handlers to import type.
