+++ src/agent/modelHandlers/openai/modelHandlerMiniMax.ts
@@ -1,5 +1,5 @@
 // Local file imports
-import { AgentWorkspaceState } from '@agent/core/execution/AgentWorkspaceState';
+import type { AgentWorkspaceState } from '@agent/core/execution/AgentWorkspaceState';
 import type { ToolDefinition } from '@model';
 import { ReasoningModelHandlerOpenAI } from './reasoningModelHandlerOpenAI';
