+++ src/agent/modelHandlers/support/MediaAttachmentProcessor.ts
@@ -7,7 +7,7 @@ import pMap from 'p-map';
 
 // Local imports - agent utils
 import { logFilesLoaded, type AgentTrace } from '@agent/trace';
-import { MediaEntry } from '@agent/utils/mediaTypes';
+import type { MediaEntry } from '@agent/utils/mediaTypes';
 import { toErrorMessage } from '@common/errors';
 import { getSdkErrorMessage } from '@common/errors/sdkErrorUtils';
 import type { FileLocation } from '@shared/schemas';
