+++ src/agent/modelHandlers/types/IModelHandler.ts
@@ -9,14 +9,14 @@ import type {
   ConversationRoundStateSnapshot,
   AgentRunStateSnapshot,
 } from '@agent/core/execution/AgentState';
-import { AgentWorkspaceState } from '@agent/core/execution/AgentWorkspaceState';
+import type { AgentWorkspaceState } from '@agent/core/execution/AgentWorkspaceState';
 import type { ProviderUsage } from '@agent/core/usage/ResponseUsage';
 import type { NormalizedUsage } from '@agent/types/NormalizedUsage';
 import type { MediaEntry } from '@agent/utils/mediaTypes';
 import type { ToolResultPayload } from '@agent/modelHandlers/utils/toolAttachmentUtils';
 import type { ToolDefinition } from '@model';
-import type { ToolFileAttachment } from '@tools/result';
 import type { FileLocation } from '@shared/schemas';
+import type { ToolFileAttachment } from '@tools/result';
 import type { ModelConfig, ModelCapabilities } from 'llm-zoo';
 import type { ProviderMessage } from './ProviderMessage';
 import type { ProviderStopReason } from './StopReasonTypes';
